Stay DRY by reading packages from turbo-install package
