+++ cli/npm/turbo-install/package.json
@@ -5,6 +5,7 @@
   "repository": "https://github.com/vercel/turborepo",
   "bugs": "https://github.com/vercel/turborepo/issues",
   "homepage": "https://turborepo.org",
+  "license": "MPL-2.0",
   "scripts": {
     "postinstall": "node install.js"
   },
@@ -16,7 +17,6 @@
     "node-platform.js",
     "install.js"
   ],
-  "license": "MPL-2.0",
   "optionalDependencies": {
     "turbo-darwin-64": "1.0.20-canary.1",
     "turbo-darwin-arm64": "1.0.20-canary.1",
